Fix to allow using OS-provided threadpool
