+++ src/coreclr/src/tools/Common/TypeSystem/Common/TypeSystemHelpers.cs
@@ -440,7 +440,7 @@ namespace Internal.TypeSystem
         }
 
         /// <summary>
-        /// Determines whether an object of type '<paramref name="type"/>' requires 8-byte alignment on 
+        /// Determines whether an object of type '<paramref name="type"/>' requires 8-byte alignment on
         /// 32bit ARM or 32bit Wasm architectures.
         /// </summary>
         public static bool RequiresAlign8(this TypeDesc type)
