+++ src/libraries/System.Private.CoreLib/src/System/Threading/LowLevelLock.cs
@@ -53,6 +53,18 @@ namespace System.Threading
             GC.SuppressFinalize(this);
         }
 
+#if DEBUG
+        public bool IsLocked
+        {
+            get
+            {
+                bool isLocked = _ownerThread == Thread.CurrentThread;
+                Debug.Assert(!isLocked || (_state & LockedMask) != 0);
+                return isLocked;
+            }
+        }
+#endif
+
         [Conditional("DEBUG")]
         public void VerifyIsLocked()
         {
